+++ config.json
@@ -1,0 +1,14 @@
+{
+    "baseAudio": {
+        "requiresOnRunClean": false
+    },
+    "wavAudio": {
+        "requiresOnRunClean": true
+    },
+    "processedAudio": {
+        "requiresOnRunClean": true
+    },
+    "streamAudio": {
+        "requiresOnRunClean": true
+    }
+}
